Improve settings, add neural net
The settings have been improved, and the danger level string now works
for the grid setup (README is yet to be updated with the format for the
string). There have been miscellaneous other changes as well: The
AdvarsarialCoverageSettings class was moved to a new file, and a
GridNodeGenerator class was created to ease grid creation. Also,
statistics functionality has been added to ease comparisons of coverage
algorithms.

Finally, a neural net implementation has been added in NeuralNet.java,
in preparation for a deep reenforcement learning coverage algorithm.