Fix minibatch trainer bug
In this commit, a bug was fixed that would cause the minibatch trainer
to learn from a random action instead of the best action.

Also, the deep q-learning code was refactored in preparation for
migration to Torch, and the ExternalTorchNN was refactored to fix a bug
that would cause the Torch script to receive SIGPIPE when output was
being retrieved.