Expand grid generator functionality
The grid generator string syntax has been expanded to include multiple
maps that are chosen at random. If the @lastInMap tag is added to a
node, the next node will be interpreted as part of a new map.

Settings can now also be changed in headless mode using the ":set"
command, and settings can be viewed with the ":showsettings" command.
Quotes and basic escape characters ('\n', '\t') work as well.