Add support for loading rc files to run commands at startup
Also added some new commands, such as ":setdisplay" which takes
arguments "gui" or "none" to control headless mode (and added a similar
option to the GUIDisplay), and ":setecho" which can be "on" or "off" to
echo back every command entered.

There is also now always a command line controller, even in GUI mode.