Fix infinite loop on external q-learner crash
This commit patches a bug that caused the simulator to continue running
when the pipe for an external q-learner was broken. Now, a broken pipe
will cause a pause of several seconds to avoid spamming the log file.
There were also numerous other changes, including:

- Refactored statistics classes into a new package
(adversarialcoverage.stats)
- Added new statistics classes to ease management of stats that are
based on samples having a sum, mean, and standard deviation.
- Shortened the names of methods in AdversarialCoverageSettings to make
code more readable.
- Added more statistics to the end-of-run and end-of-batch outputs.